fix: rename to useLensAuthData
