+++ apps/web/src/components/Common/Providers/ProProvider.tsx
@@ -3,7 +3,7 @@ import type { FC } from 'react';
 import { STALE_TIMES } from '@hey/data/constants';
 import getPro from '@hey/helpers/api/getPro';
 import { useQuery } from '@tanstack/react-query';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { useProStore } from 'src/store/non-persisted/useProStore';
 
 const ProProvider: FC = () => {
