+++ apps/web/src/components/Composer/Post/Discard.tsx
@@ -4,7 +4,7 @@ import { HEY_API_URL } from '@hey/data/constants';
 import { Alert } from '@hey/ui';
 import axios from 'axios';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { useCollectModuleStore } from 'src/store/non-persisted/publication/useCollectModuleStore';
 import { usePublicationStore } from 'src/store/non-persisted/publication/usePublicationStore';
 import { useGlobalModalStateStore } from 'src/store/non-persisted/useGlobalModalStateStore';
