+++ apps/web/src/components/Profile/CreatorTool.tsx
@@ -16,7 +16,7 @@ import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 
 interface CreatorToolProps {
   profile: Profile;
