+++ apps/web/src/components/Publication/Actions/Tip/Action.tsx
@@ -22,8 +22,8 @@ import cn from '@hey/ui/cn';
 import axios from 'axios';
 import { useRef, useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
 import useHandleWrongNetwork from 'src/hooks/useHandleWrongNetwork';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import usePreventScrollOnNumberInput from 'src/hooks/usePreventScrollOnNumberInput';
 import { useGlobalModalStateStore } from 'src/store/non-persisted/useGlobalModalStateStore';
 import { useProfileStatus } from 'src/store/non-persisted/useProfileStatus';
