+++ apps/web/src/components/Publication/OpenAction/CollectModule/CollectAction.tsx
@@ -35,8 +35,8 @@ import { Button, Spinner, WarningMessage } from '@hey/ui';
 import cn from '@hey/ui/cn';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
 import useHandleWrongNetwork from 'src/hooks/useHandleWrongNetwork';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { useNonceStore } from 'src/store/non-persisted/useNonceStore';
 import { useProfileStatus } from 'src/store/non-persisted/useProfileStatus';
 import { useProfileStore } from 'src/store/persisted/useProfileStore';
