+++ apps/web/src/components/Publication/OpenAction/UnknownModule/ActionButton.tsx
@@ -12,7 +12,7 @@ import { Button, Spinner, WarningMessage } from '@hey/ui';
 import cn from '@hey/ui/cn';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { formatUnits } from 'viem';
 import { useAccount, useBalance } from 'wagmi';
