+++ apps/web/src/components/Settings/Pro/AppIcon.tsx
@@ -9,7 +9,7 @@ import { Card, CardHeader, Tooltip } from '@hey/ui';
 import axios from 'axios';
 import { useState } from 'react';
 import { toast } from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { usePreferencesStore } from 'src/store/non-persisted/usePreferencesStore';
 
 const icons = [
