+++ apps/web/src/components/Shared/Club/Leave.tsx
@@ -6,7 +6,7 @@ import { Button } from '@hey/ui';
 import axios from 'axios';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 
 interface LeaveProps {
   id: string;
