+++ apps/web/src/components/Shared/Oembed/Frames/Transaction.tsx
@@ -12,7 +12,7 @@ import { Button } from '@hey/ui';
 import axios from 'axios';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { useProfileStore } from 'src/store/persisted/useProfileStore';
 import {
   arbitrum,
