+++ apps/web/src/components/Staff/FeatureFlags/Create.tsx
@@ -8,7 +8,7 @@ import { Button, Form, Input, useZodForm } from '@hey/ui';
 import axios from 'axios';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 import { object, string } from 'zod';
 
 const createFeatureSchema = object({
