+++ apps/web/src/components/Staff/FeatureFlags/List.tsx
@@ -16,7 +16,7 @@ import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 
 import Assign from './Assign';
 import Create from './Create';
