+++ apps/web/src/components/Staff/Overview/LeafwatchStats.tsx
@@ -11,7 +11,7 @@ import { CardHeader, ErrorMessage, NumberedStat } from '@hey/ui';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { useState } from 'react';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 
 import ActiveUsers from './ActiveUsers';
 import EventsToday from './EventsToday';
