+++ apps/web/src/components/Staff/Users/Overview/Tool/ProfileOverview.tsx
@@ -18,7 +18,7 @@ import { APP_NAME, HEY_API_URL } from '@hey/data/constants';
 import formatAddress from '@hey/helpers/formatAddress';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 
 interface ProfileOverviewProps {
   profile: Profile;
