+++ apps/web/src/components/Staff/Users/Overview/Tool/UpdateFeatureFlags.tsx
@@ -11,7 +11,7 @@ import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { useState } from 'react';
 import toast from 'react-hot-toast';
-import useLensAuthData from 'src/hooks/useAuthApiHeaders';
+import useLensAuthData from 'src/hooks/useLensAuthData';
 
 import ToggleWrapper from './ToggleWrapper';
