+++ apps/web/src/hooks/useCreatePoll.tsx
@@ -2,7 +2,7 @@ import { HEY_API_URL } from '@hey/data/constants';
 import axios from 'axios';
 import { usePublicationPollStore } from 'src/store/non-persisted/publication/usePublicationPollStore';
 
-import useLensAuthData from './useAuthApiHeaders';
+import useLensAuthData from './useLensAuthData';
 
 type CreatePollResponse = string;
